# RestApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Superhero API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/getuser": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "7508d38362951e3be4c31513644ab5e4",
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No user found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No user found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/connect": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Connect a user",
                "operationId": "84850680513ee0394084746688ec0724",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of the user",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password matched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password matched"
                                        },
                                        "cookie": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Password mismatch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password mismatch"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/createuser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a user",
                "operationId": "ba820f5ee59d44aecf828e317515bc95",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "Firstname of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "remember_token",
                        "in": "query",
                        "description": "Remember token of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Api token of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "Avatar of the user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "example of the body request",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "banos"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "b@.Fr"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "flodagnas54"
                                    },
                                    "remember_token": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "api_token": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/checkSession": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Check if the session is valid",
                "operationId": "efea466fd2a8c299b565211de8b8db09",
                "parameters": [
                    {
                        "name": "cookie",
                        "in": "query",
                        "description": "Cookie of the session",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Session valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Session valid"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Session not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "400"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Session not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}